+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Master
+## 0.30.1: Localized Stain Remover
 
 #### Breaking